requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
google-api-core